requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
pypdf
docx2txt
chromadb
filetype
python-dotenv
streamlit

